Move "Docopt.CreateParser" to "Docopt.cs"
