+++ src/DocoptNet/ParseResult.cs
@@ -3,7 +3,6 @@
 namespace DocoptNet
 {
     using System;
-    using System.Collections.Generic;
 
     partial interface IArgumentsResult<out T>
     {
@@ -156,26 +155,5 @@ namespace DocoptNet
         TResult IBaselineParser<T>.IResult.Match<TResult>(Func<T, TResult> args,
                                                           Func<IInputErrorResult, TResult> error) =>
             error(this);
-    }
-
-    partial class Docopt
-    {
-        public static IHelpFeaturingParser<IDictionary<string, Value>> CreateParser(string doc) =>
-            new Parser<IDictionary<string, Value>>(doc, static (doc, argv, flags, version) =>
-            {
-                var optionsFirst = (flags & ParseFlags.OptionsFirst) != ParseFlags.None;
-                var parsedResult = Parse(doc, Tokens.From(argv), optionsFirst);
-
-                var help = (flags & ParseFlags.DisableHelp) == ParseFlags.None;
-                if (help && parsedResult.IsHelpOptionSpecified)
-                    return new ParseHelpResult<IDictionary<string, Value>>(doc);
-
-                if (version is { } someVersion && parsedResult.IsVersionOptionSpecified)
-                    return new ParseVersionResult<IDictionary<string, Value>>(someVersion);
-
-                return parsedResult.TryApply(out var applicationResult)
-                     ? new ArgumentsResult<IDictionary<string, Value>>(applicationResult.ToValueDictionary())
-                     : new ParseInputErrorResult<IDictionary<string, Value>>("Input error.", parsedResult.ExitUsage);
-            });
     }
 }
